Make DatastoreFile work with an actual Datastore
There were a number of issues with saving entities represented by
DatastoreFile objects which have been resolved. For instance it is no
longer possible to create a DatastoreFile with a null Array of Bytes.
The contents property of Entity objects has also been changed to be
stored as a Blob rather than as an Array of Bytes because the App Engine
Datastore will not allow a bare Array.