Convert DatastoreFile to use List internnally
List is easier to work with in many cases and so it makes sense to use
it as the default structure for holding a sequence of data. Additionally
this allows constructing files with zero length which makes more sense
than constructing empty files with a default length of 32 bytes.
DatastoreFile::readBytes now also advances position and returns a
DatastoreFile instance which DatastoreInput stores.
DatastoreFile::writeBytes now advances position according to length of
bytes written.