Correct behavior of DatastoreFile read and write methods
If reading more bytes than exist in the file throw an IOException noting
we are reading past the end of the file.

When writing an array of bytes only write the length of bytes specified
in the len parameter. Additionally the length of bytes to write starts
at the offset in the array of bytes to write and writes starting at the
current position.