Add task queuing to the QuoteServlet
In order to do this, an abstract enqueue method was added to
QuoteController, it is implemented in DatastoreQuoteController. In order
to continue being able to unit test or QuoteController functionality
save had to be split out so it just did the Persister.save call.
Indexing is in a separate method. QuoteController::saveAndIndex was
added in order to be able to test without an active web server. This is
the cause of separate save and enqueue calls in QuoteServlet.