Add a lot more logic to QcServlet
Logic includes an inner HttpHelper class which is sort of set up to
handle rest like requests. QcServlet also can handle most servlet
requests so the only thing which must be defined in the servlet is the
context to render the view with and the name/path of the view.